Create OrderController

+++ backend-spring-boot/Online Food Ordering/Online-Food-Ordering/src/main/java/com/heli/model/Order.java
@@ -55,6 +55,6 @@ public class Order {
     private int totalItem;
 
     @Column(name = "total_price")
-    private int totalPrice;
+    private Long totalPrice;
 
 }
